requirements.txt
cassandra-driver~=3.25.0

redis~=4.0.2